requirements.txt
fastapi
uvicorn
pandas
scikit-learn
sentence-transformers
supabase
python-dotenv